moved districtyears and cardcreator into folders as well as split up their css and create test files for them
