+++ src/App/App.css
@@ -1,4 +1,4 @@
 body {
-  background-color: #B8D5B8;
-  color: #383F51;
+  background-color: #333745;
+  color: #333745;
 }
