+++ src/DistrictCards/DistrictCards.css
@@ -16,7 +16,7 @@ beige: #eef5db*/
 .district-card {
   margin: 5px;
   border-radius: 5px;
-  border: 5px solid #383F51;
+  border: 5px solid #333745;
   max-width: 300px;
   min-width: 300px;
   background-color: #FFF;
@@ -26,13 +26,4 @@ beige: #eef5db*/
   padding: 0px 10px;
   text-align: center;
   font-size: 20px;
-}
-
-.yearly-data {
-  padding: 0% 0% 3% 35%
-}
-
-.over {
-  font-weight: bolder;
-  text-decoration:underline;
 }
