+++ src/DistrictYears/DistrictYears.css
@@ -1,0 +1,8 @@
+.yearly-data {
+  padding: 0% 0% 3% 35%
+}
+
+.over {
+  font-weight: bolder;
+  text-decoration:underline;
+}
